adding filters to the system builder index
Rendered partials, are now incorporated to the build index trying to code the logic

when user/builder picks a specific component all the non compatible parts are hidden.

logically: we can gsub to remove the values that are not important to the  builder (i.e sockets, we'll be limiting cpu sockets to AM4 and LGA1151 / LGA1200 )

This commit is an update to our WIP branch so the team can chip in their idea with the filters

Lightbulb moment :
I want to render our partials on the div "on button click" if the team has an idea how these can be easily done let me know.